Add files via upload
Write a program that reads two char matrices (A[][] and B[][]) of the same order M * N and prints the third matrix C[][] which is filled with the intersecting elements of A and B, otherwise set the element to '*'. On the first two lines, you receive M and N, then on 2 * M lines N characters – the matrices elements.
The matrix elements may be any ASCII char except '*'.