Add files via upload
Write a program that reads a matrix from the console. Then find the biggest sum of a 2x2 submatrix. Print the submatrix and its sum.
On the first line, you will get the dimensions of the matrix in the format "{rows, columns}". On the next lines, you will get the elements for each row separated by a comma.