Add files via upload
Write a program that reads a matrix from the console. Then print the diagonals. The matrix will always be square. On the first line, you read a single integer N the matrix size. Then on each line N elements. The first diagonal should always start with the element at the first row and col, the second diagonal should start with the element at the last row and first col.