Add files via upload
You will be given the rows of a matrix. Then the matrix itself. Inside this matrix, there are mistaken values that need to be replaced you will receive the wrong value at the last line. Those values should be replaced with the sum of the nearest elements in the four directions up, down, left, and right, but only if they are valid values. In the end, you have to print the fixed measurements.